Comments added/changed
A notable bug was discovered and temporarily addressed- an attempt to remove the hardcoded ID number by converting the uint8_t value to a String for comparison caused the Feather M0 to crash. This forced the use of the bootloader mode in order to reprogram. A note has been made and the offending lines have been reverted.